+++ tests/test_core/test_ws.py
@@ -1,9 +1,14 @@
-# from lihil.websocket import WebSocketRoute
-# from lihil import LocalClient
+from lihil.websocket import WebSocketRoute, WebSocket
+from lihil import LocalClient, Lihil
 
 
 
-# async def test_ws():
-#     lc = LocalClient()
+async def test_ws():
+    lc = LocalClient()
 
-#     ws = WebSocketRoute("test")
+    ws = WebSocketRoute("test")
+    async def test_ws(ws: WebSocket):
+        await ws.accept()
+
+    lhl = Lihil()
+    lhl.include_routes(ws)