+++ CHANGELOG.md
@@ -845,4 +845,14 @@ async def get_user(
 ```
 
 
+- [ ] websocket
+
+```python
+ async def websocket_endpoint(websocket: WebSocket) -> None:
+    await websocket.accept()
+    while True:
+        data = await websocket.receive_text()
+        await websocket.send_text(f"Message text was: {data}")
+```
+
 - [ ] static response